backend/requirements.txt
# Kindra CBO Backend - Production Requirements

# Core Django & REST Framework
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Caching & Task Queue
redis==5.0.1
celery==5.3.6
django-celery-beat==2.5.0
django-celery-results==2.5.1

# File Storage
cloudinary==1.40.0
django-cloudinary-storage==0.3.0

# Environment Variables
python-decouple==3.8

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0

# Monitoring & Error Tracking
sentry-sdk==1.40.0

# Email
django-sendgrid-v5==1.2.3

# API Documentation
drf-spectacular==0.27.1

# Utilities
Pillow==10.2.0
python-dateutil==2.8.2
pytz==2024.1

# Security
django-ratelimit==4.1.0